+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## v1.0.2 - 2024/02/15
  - Implemented increment/decrement buttons for Phase times, long overdue QOL feature.
+ - Fixed a bug when omitting scan time in sample-sequence CSVs.
+ - Fixed a logic error on quit when log backup location can't be found.
 
 ## v1.0.1 - 2024/02/14
  - Fixed various bugs left over in v1.0.0: